Improving code quality, splitting components.
