+++ web/src/components/column-level/ColumnLevel.tsx
@@ -1,21 +1,20 @@
 import * as Redux from 'redux'
-import { ArrowBackIosRounded, CropFree, Refresh, ZoomIn, ZoomOut } from '@mui/icons-material'
+import { ActionBar } from './ActionBar'
 import { ColumnLineageGraph } from '../../types/api'
-import { Divider, Drawer, TextField, Tooltip } from '@mui/material'
+import { Drawer } from '@mui/material'
 import { Graph, ZoomPanControls } from '../../../libs/graph'
 import { IState } from '../../store/reducers'
 import { MultipleNodeData, MultipleNodeKind, columnLevelNodeRenderer } from './nodes'
+import { ZoomControls } from './ZoomControls'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 import { createElkNodes, parseColumnLineageNode } from './layout'
 import { fetchColumnLineage } from '../../store/actionCreators'
 import { theme } from '../../helpers/theme'
 import { useCallbackRef } from '../../helpers/hooks'
-import { useNavigate, useParams, useSearchParams } from 'react-router-dom'
+import { useParams, useSearchParams } from 'react-router-dom'
 import Box from '@mui/material/Box'
 import ColumnLevelDrawer from './ColumnLevelDrawer'
-import IconButton from '@mui/material/IconButton'
-import MqText from '../core/text/MqText'
 import ParentSize from '@visx/responsive/lib/components/ParentSize'
 import React, { useEffect, useRef, useState } from 'react'
 
@@ -37,7 +36,6 @@ const ColumnLevel: React.FC<ColumnLevelProps> = ({
   columnLineage: columnLineage,
 }: ColumnLevelProps) => {
   const { namespace, name } = useParams()
-  const navigate = useNavigate()
   const [searchParams, setSearchParams] = useSearchParams()
 
   const [depth, setDepth] = useState(Number(searchParams.get('depth')) || 2)
@@ -85,70 +83,7 @@ const ColumnLevel: React.FC<ColumnLevelProps> = ({
 
   return (
     <>
-      <Box
-        sx={{
-          borderBottomWidth: 2,
-          borderTopWidth: 0,
-          borderLeftWidth: 0,
-          borderRightWidth: 0,
-          borderStyle: 'dashed',
-        }}
-        display={'flex'}
-        height={'64px'}
-        justifyContent={'space-between'}
-        alignItems={'center'}
-        px={2}
-        borderColor={theme.palette.secondary.main}
-      >
-        <Box display={'flex'} alignItems={'center'}>
-          <Tooltip title={'Back to datasets'}>
-            <IconButton size={'small'} sx={{ mr: 2 }} onClick={() => navigate('/datasets')}>
-              <ArrowBackIosRounded fontSize={'small'} />
-            </IconButton>
-          </Tooltip>
-          <MqText heading>Datasets</MqText>
-          <Divider orientation='vertical' flexItem sx={{ mx: 2 }} />
-          <Box>
-            <MqText subdued>Namespace</MqText>
-            <MqText font={'mono'}>{namespace || 'Unknown namespace name'}</MqText>
-          </Box>
-          <Divider orientation='vertical' flexItem sx={{ mx: 2 }} />
-          <Box>
-            <MqText subdued>Name</MqText>
-            <MqText font={'mono'}>{name || 'Unknown dataset name'}</MqText>
-          </Box>
-        </Box>
-        <Box display={'flex'} alignItems={'center'}>
-          <Tooltip title={'Refesh'}>
-            <IconButton
-              sx={{ mr: 2 }}
-              color={'primary'}
-              size={'small'}
-              onClick={() => {
-                if (namespace && name) {
-                  fetchColumnLineage('DATASET', namespace, name, depth)
-                }
-              }}
-            >
-              <Refresh fontSize={'small'} />
-            </IconButton>
-          </Tooltip>
-          <TextField
-            id='column-level-depth'
-            type='number'
-            label='Depth'
-            variant='outlined'
-            size='small'
-            sx={{ width: '80px' }}
-            value={depth}
-            onChange={(e) => {
-              setDepth(isNaN(parseInt(e.target.value)) ? 0 : parseInt(e.target.value))
-              searchParams.set('depth', e.target.value)
-              setSearchParams(searchParams)
-            }}
-          />
-        </Box>
-      </Box>
+      <ActionBar fetchColumnLineage={fetchColumnLineage} depth={depth} setDepth={setDepth} />
       <Box height={'calc(100vh - 98px - 64px)'}>
         <Drawer
           anchor={'right'}
@@ -159,33 +94,7 @@ const ColumnLevel: React.FC<ColumnLevelProps> = ({
             <ColumnLevelDrawer />
           </Box>
         </Drawer>
-        <Box
-          display={'flex'}
-          border={1}
-          borderRadius={1}
-          flexDirection={'column'}
-          m={1}
-          position={'absolute'}
-          right={0}
-          zIndex={1}
-          borderColor={theme.palette.grey[500]}
-        >
-          <Tooltip title={'Zoom in'} placement={'left'}>
-            <IconButton size='small' onClick={() => handleScaleZoom('in')}>
-              <ZoomIn />
-            </IconButton>
-          </Tooltip>
-          <Tooltip title={'Zoom out'} placement={'left'}>
-            <IconButton size='small' onClick={() => handleScaleZoom('out')}>
-              <ZoomOut />
-            </IconButton>
-          </Tooltip>
-          <Tooltip title={'Reset zoom'} placement={'left'}>
-            <IconButton size={'small'} onClick={handleResetZoom}>
-              <CropFree />
-            </IconButton>
-          </Tooltip>
-        </Box>
+        <ZoomControls handleScaleZoom={handleScaleZoom} handleResetZoom={handleResetZoom} />
         <ParentSize>
           {(parent) => (
             <Graph<MultipleNodeKind, MultipleNodeData>
